Refactor application into package structure
